requirements.txt
fastapi
uvicorn
streamlit
python-dotenv
google-generativeai
pydantic
requests
